chore(migrations): remove unused test migration file
The test migration file '548795eaa1aa_create_test_migration.py' was deleted as it was no longer needed. This file contained auto-generated Alembic commands to create tables for roles, users, addresses, preferences, and user_roles. The downgrade function was also provided to drop these tables if needed. Since this migration was not used in the project, it was decided to remove it to keep the migrations directory clean and avoid confusion.